refactor(quality): split UserProfile, optimize transit perf, semantic HTML, file guidelines
UserProfile Split (#218):
- Split 1084-line UserProfile.tsx into 6 focused sub-components
  - ProfileHeader (237 lines): avatar, user info, edit mode
  - AccountTab (96 lines): account details, security, danger zone
  - ChartsTab (130 lines): charts list with grid
  - PreferencesTab (180 lines): house system, zodiac, theme
  - SubscriptionTab (157 lines): plans, billing
  - types.ts (83 lines): shared interfaces
- UserProfile.tsx reduced to ~237-line orchestrator
- All existing imports remain backward-compatible

Transit Performance (#274):
- Add sampling intervals for long-duration transit forecasts
- year: sample every 7 days (52 iterations vs 365)
- quarter: sample every 3 days (30 iterations vs 90)
- week/month: daily as before

Semantic HTML (#131):
- HomePage: add aria-labels to 5 existing sections
- SavedChartsGalleryPage: div → section with aria-label for search + gallery

Documentation (#127):
- Add file size guidelines table to CLAUDE.md
- Document extraction procedure and enforcement rules
