+++ frontend/src/components/UserProfile.tsx
@@ -1,94 +1,28 @@
-import { useState, useCallback } from 'react';
-import { useNavigate } from 'react-router-dom';
+import { useState } from 'react';
 import { useAuth, useCharts } from '../hooks';
-import { EmptyState } from './EmptyState';
 import { logger } from '@/utils/logger';
-import locationService from '@/services/location.service';
-import { APP_LOCALE } from '../utils/constants';
+import { ProfileHeader } from './user-profile/ProfileHeader';
+import { AccountTab } from './user-profile/AccountTab';
+import { ChartsTab } from './user-profile/ChartsTab';
+import { PreferencesTab } from './user-profile/PreferencesTab';
+import { SubscriptionTab } from './user-profile/SubscriptionTab';
+import type { UserProfile, Chart, UserData, UserProfileProps } from './user-profile/types';
 
-// Types based on findings.md
-export interface UserProfile {
-  id: string;
-  email: string;
-  name: string;
-  avatar?: string;
-  timezone: string;
-  createdAt: Date;
-  subscription: {
-    plan: 'free' | 'premium' | 'professional';
-    status: 'active' | 'canceled' | 'expired';
-    renewalDate?: Date;
-  };
-  preferences: {
-    theme: 'light' | 'dark' | 'auto';
-    defaultHouseSystem: HouseSystem;
-    defaultZodiac: 'tropical' | 'sidereal';
-    aspectOrbs: {
-      conjunction: number;
-      opposition: number;
-      trine: number;
-      square: number;
-      sextile: number;
-    };
-  };
+// Re-export types for backward compatibility (barrel imports from components/index.ts)
+export type { UserProfile as UserProfileType, Chart, HouseSystem } from './user-profile/types';
+
+interface TabConfig {
+  id: 'account' | 'charts' | 'preferences' | 'subscription';
+  label: string;
+  icon: string;
 }
 
-export type HouseSystem = 'placidus' | 'koch' | 'porphyry' | 'whole' | 'equal' | 'topocentric';
-
-export interface Chart {
-  id: string;
-  userId: string;
-  name: string;
-  type: 'natal' | 'synastry' | 'composite' | 'transit';
-  birthData: {
-    date: Date;
-    time: string;
-    place: {
-      name: string;
-      latitude: number;
-      longitude: number;
-      timezone: string;
-    };
-    timeUnknown: boolean;
-  };
-  settings: {
-    houseSystem: HouseSystem;
-    zodiac: 'tropical' | 'sidereal';
-    sideralMode?: string;
-  };
-  calculatedData?: Record<string, unknown>;
-  createdAt: Date;
-  updatedAt: Date;
-}
-
-interface UserProfileProps {
-  onEditChart?: (chartId: string) => void;
-  onViewChart?: (chartId: string) => void;
-  onDeleteChart?: (chartId: string) => void;
-}
-
-interface UserData {
-  id: string;
-  email: string;
-  name: string;
-  avatar_url?: string;
-  avatar?: string;
-  createdAt?: Date;
-  timezone?: string;
-  plan?: string;
-  preferences?: {
-    theme?: 'light' | 'dark' | 'auto';
-    defaultHouseSystem?: HouseSystem;
-    defaultZodiac?: 'tropical' | 'sidereal';
-    aspectOrbs?: {
-      conjunction: number;
-      opposition: number;
-      trine: number;
-      square: number;
-      sextile: number;
-    };
-  };
-}
+const TABS: TabConfig[] = [
+  { id: 'account', label: 'Account', icon: '👤' },
+  { id: 'charts', label: 'My Charts', icon: '📊' },
+  { id: 'preferences', label: 'Preferences', icon: '⚙️' },
+  { id: 'subscription', label: 'Subscription', icon: '💳' },
+];
 
 export function UserProfile({ onEditChart, onViewChart, onDeleteChart }: UserProfileProps) {
   const { user, updateProfile } = useAuth();
@@ -171,13 +105,6 @@ export function UserProfile({ onEditChart, onViewChart, onDeleteChart }: UserPro
     timezone: user?.timezone ?? 'UTC',
   });
 
-  const tabs = [
-    { id: 'account' as const, label: 'Account', icon: '👤' },
-    { id: 'charts' as const, label: 'My Charts', icon: '📊' },
-    { id: 'preferences' as const, label: 'Preferences', icon: '⚙️' },
-    { id: 'subscription' as const, label: 'Subscription', icon: '💳' },
-  ];
-
   const handleSaveProfile = (): void => {
     void (async () => {
       try {
@@ -223,12 +150,12 @@ export function UserProfile({ onEditChart, onViewChart, onDeleteChart }: UserPro
             aria-label="Profile sections"
             className="flex overflow-x-auto"
             onKeyDown={(e) => {
-              const idx = tabs.findIndex((t) => t.id === activeTab);
-              if (e.key === 'ArrowRight' && idx < tabs.length - 1) { e.preventDefault(); setActiveTab(tabs[idx + 1].id); }
-              if (e.key === 'ArrowLeft' && idx > 0) { e.preventDefault(); setActiveTab(tabs[idx - 1].id); }
+              const idx = TABS.findIndex((t) => t.id === activeTab);
+              if (e.key === 'ArrowRight' && idx < TABS.length - 1) { e.preventDefault(); setActiveTab(TABS[idx + 1].id); }
+              if (e.key === 'ArrowLeft' && idx > 0) { e.preventDefault(); setActiveTab(TABS[idx - 1].id); }
             }}
           >
-            {tabs.map((tab) => (
+            {TABS.map((tab) => (
               <button
                 type="button"
                 key={tab.id}
@@ -305,780 +232,6 @@ export function UserProfile({ onEditChart, onViewChart, onDeleteChart }: UserPro
           </div>
         </div>
       )}
-    </div>
-  );
-}
-
-// Common IANA timezones for the fallback manual select
-const COMMON_TIMEZONES = [
-  { value: 'Pacific/Honolulu', label: 'Hawaii (UTC-10)' },
-  { value: 'America/Anchorage', label: 'Alaska (UTC-9)' },
-  { value: 'America/Los_Angeles', label: 'Pacific Time (UTC-8)' },
-  { value: 'America/Denver', label: 'Mountain Time (UTC-7)' },
-  { value: 'America/Chicago', label: 'Central Time (UTC-6)' },
-  { value: 'America/New_York', label: 'Eastern Time (UTC-5)' },
-  { value: 'America/Sao_Paulo', label: 'São Paulo (UTC-3)' },
-  { value: 'Atlantic/Reykjavik', label: 'Iceland (UTC+0)' },
-  { value: 'Europe/London', label: 'London (UTC+0/+1)' },
-  { value: 'Europe/Paris', label: 'Paris / CET (UTC+1)' },
-  { value: 'Europe/Istanbul', label: 'Istanbul (UTC+3)' },
-  { value: 'Europe/Moscow', label: 'Moscow (UTC+3)' },
-  { value: 'Asia/Dubai', label: 'Dubai (UTC+4)' },
-  { value: 'Asia/Karachi', label: 'Karachi (UTC+5)' },
-  { value: 'Asia/Kolkata', label: 'India (UTC+5:30)' },
-  { value: 'Asia/Dhaka', label: 'Dhaka (UTC+6)' },
-  { value: 'Asia/Bangkok', label: 'Bangkok (UTC+7)' },
-  { value: 'Asia/Shanghai', label: 'China (UTC+8)' },
-  { value: 'Asia/Singapore', label: 'Singapore (UTC+8)' },
-  { value: 'Asia/Tokyo', label: 'Japan (UTC+9)' },
-  { value: 'Asia/Seoul', label: 'Korea (UTC+9)' },
-  { value: 'Australia/Sydney', label: 'Sydney (UTC+10/+11)' },
-  { value: 'Pacific/Auckland', label: 'Auckland (UTC+12/+13)' },
-  { value: 'UTC', label: 'UTC' },
-];
-
-// Profile Header Component
-function ProfileHeader({
-  user,
-  isEditing,
-  editData,
-  onEditToggle,
-  onDataChange,
-  onSave,
-  birthCoordinates,
-}: {
-  user?: UserProfile;
-  isEditing: boolean;
-  editData: { name: string; timezone: string };
-  onEditToggle: () => void;
-  onDataChange: (data: { name: string; timezone: string }) => void;
-  onSave: () => void;
-  birthCoordinates?: { latitude: number; longitude: number } | null;
-}) {
-  const [detectingTimezone, setDetectingTimezone] = useState(false);
-  const [detectedTimezone, setDetectedTimezone] = useState<string | null>(null);
-  const [detectionError, setDetectionError] = useState<string | null>(null);
-
-  const handleDetectTimezone = useCallback(async () => {
-    if (!birthCoordinates) return;
-    setDetectingTimezone(true);
-    setDetectionError(null);
-    try {
-      const tz = await locationService.getTimezone(
-        birthCoordinates.latitude,
-        birthCoordinates.longitude,
-      );
-      setDetectedTimezone(tz);
-      onDataChange({ ...editData, timezone: tz });
-    } catch {
-      setDetectionError('Could not detect timezone from birth location.');
-    } finally {
-      setDetectingTimezone(false);
-    }
-  }, [birthCoordinates, editData, onDataChange]);
-
-  if (!user) return null;
-
-  return (
-    <div className="glass-panel rounded-2xl relative overflow-hidden p-6 text-white">
-      {/* Background Decoration */}
-      <div className="absolute -top-16 -right-16 w-48 h-48 bg-primary/20 rounded-full blur-[80px] pointer-events-none" />
-      <div className="absolute -bottom-16 -left-16 w-48 h-48 bg-cosmic-blue/10 rounded-full blur-[80px] pointer-events-none" />
-      <div className="flex flex-col sm:flex-row items-center gap-6">
-        {/* Avatar */}
-        <div className="relative group">
-          <div className="w-24 h-24 rounded-full p-[3px] bg-gradient-to-tr from-primary via-cosmic-blue to-primary/30 relative">
-            <div className="w-full h-full rounded-full bg-cosmic-page p-1">
-              <div className="w-full h-full rounded-full bg-cosmic-card-solid flex items-center justify-center text-3xl font-bold text-white">
-                {user.name.charAt(0).toUpperCase()}
-              </div>
-            </div>
-          </div>
-          <button
-            type="button"
-            className="absolute bottom-0 right-0 p-2 bg-white rounded-full shadow-lg opacity-0 group-hover:opacity-100 transition-opacity"
-            aria-label="Change avatar"
-            title="Change avatar"
-          >
-            <span className="material-symbols-outlined text-4xl text-primary" aria-hidden="true">photo_camera</span>
-          </button>
-        </div>
-
-        {/* User Info */}
-        <div className="flex-1 text-center sm:text-left">
-          {isEditing ? (
-            <div className="space-y-3">
-              <div>
-                <label htmlFor="profile-name-edit" className="sr-only">
-                  Edit Display Name
-                </label>
-                <input
-                  type="text"
-                  id="profile-name-edit"
-                  data-testid="profile-name-edit"
-                  value={editData.name}
-                  onChange={(e) => onDataChange({ ...editData, name: e.target.value })}
-                  aria-required="true"
-                  className="px-4 py-2 rounded-xl bg-cosmic-card-solid border border-white/15 text-white placeholder:text-slate-200 focus:outline-none focus:ring-2 focus:ring-primary focus:border-primary"
-                  placeholder="Your name"
-                />
-              </div>
-
-              {/* Timezone: auto-detect + manual fallback */}
-              <div className="space-y-2">
-                {/* Detected timezone display */}
-                <div className="flex items-center gap-2 flex-wrap">
-                  <span className="text-sm text-slate-200">Timezone:</span>
-                  {editData.timezone && editData.timezone !== 'UTC' && !editData.timezone.startsWith('UTC') && (
-                    <span className="inline-flex items-center gap-1 px-2 py-1 rounded-lg bg-primary/15 text-primary text-sm font-medium">
-                      <span className="material-symbols-outlined text-4xl" aria-hidden="true">language</span>
-                      {editData.timezone}
-                    </span>
-                  )}
-                  <button
-                    type="button"
-                    onClick={() => { void handleDetectTimezone(); }}
-                    disabled={detectingTimezone || !birthCoordinates}
-                    className="inline-flex items-center gap-1 px-3 py-1.5 text-xs font-medium rounded-lg bg-white/15 border border-white/15 text-slate-200 hover:bg-white/20 hover:text-white transition-colors disabled:opacity-40 disabled:cursor-not-allowed"
-                    title={birthCoordinates ? 'Detect timezone from your birth location' : 'No birth location data available'}
-                  >
-                    <span className="material-symbols-outlined text-4xl" aria-hidden="true">
-                      {detectingTimezone ? 'hourglass_empty' : 'my_location'}
-                    </span>
-                    {detectingTimezone ? 'Detecting…' : 'Detect from birth location'}
-                  </button>
-                </div>
-
-                {detectionError && (
-                  <p className="text-xs text-red-400">{detectionError}</p>
-                )}
-                {detectedTimezone && !detectionError && (
-                  <p className="text-xs text-green-400">
-                    ✓ Detected: {detectedTimezone}
-                  </p>
-                )}
-
-                {/* Fallback manual select */}
-                <div>
-                  <label htmlFor="profile-timezone-edit" className="sr-only">
-                    Timezone
-                  </label>
-                  <select
-                    id="profile-timezone-edit"
-                    value={COMMON_TIMEZONES.some(tz => tz.value === editData.timezone) ? editData.timezone : '__custom__'}
-                    onChange={(e) => {
-                      if (e.target.value !== '__custom__') {
-                        onDataChange({ ...editData, timezone: e.target.value });
-                      }
-                    }}
-                    className="px-4 py-2 rounded-xl bg-cosmic-card-solid border border-white/15 text-white focus:outline-none focus:ring-2 focus:ring-primary focus:border-primary text-sm"
-                    style={{ colorScheme: 'dark' }}
-                  >
-                    <option value="__custom__" disabled style={{ background: '#1a1c2e', color: '#fff' }}>
-                      {editData.timezone} (current)
-                    </option>
-                    {COMMON_TIMEZONES.map((tz) => (
-                      <option key={tz.value} value={tz.value} style={{ background: '#1a1c2e', color: '#fff' }}>
-                        {tz.label}
-                      </option>
-                    ))}
-                  </select>
-                  <p className="mt-1 text-xs text-slate-200/60">
-                    Or pick manually from common timezones above
-                  </p>
-                </div>
-              </div>
-            </div>
-          ) : (
-            <div>
-              <h1 className="text-2xl font-bold gradient-text">{user.name}</h1>
-              <p className="text-slate-200">{user.email}</p>
-              <p className="text-sm text-slate-200 mt-1">
-                Timezone: <span className="text-primary">{user.timezone}</span>
-              </p>
-              <p className="text-sm text-slate-200 mt-1">
-                Member since {new Date(user.createdAt).toLocaleDateString(APP_LOCALE, { month: 'long', year: 'numeric' })}
-              </p>
-            </div>
-          )}
-        </div>
-
-        {/* Edit Button */}
-        <div className="flex gap-2">
-          {isEditing ? (
-            <>
-              <button
-                type="button"
-                onClick={() => { void onSave(); }}
-                className="px-4 py-2 bg-cosmic-gradient text-white rounded-xl font-medium shadow-lg shadow-primary/25 hover:opacity-90 transition-opacity"
-                aria-label="Save"
-              >
-                Save
-              </button>
-              <button
-                type="button"
-                onClick={onEditToggle}
-                className="p-2 bg-white/15 border border-white/15 rounded-xl hover:bg-white/15 transition-colors"
-                aria-label="Cancel"
-              >
-                <span className="material-symbols-outlined text-4xl" aria-hidden="true">close</span>
-              </button>
-            </>
-          ) : (
-            <button
-              type="button"
-              onClick={onEditToggle}
-              className="p-2 bg-white/15 border border-white/15 rounded-xl hover:bg-white/15 transition-colors"
-              title="Edit profile"
-              aria-label="Edit profile"
-            >
-              <span className="material-symbols-outlined text-4xl" aria-hidden="true">edit</span>
-            </button>
-          )}
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// Account Tab Component
-function AccountTab({ user }: { user?: UserProfile }) {
-  if (!user) return null;
-
-  return (
-    <div className="space-y-6">
-      <h3 className="text-lg font-semibold text-white">Account Details</h3>
-
-      <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-        <div>
-          <label htmlFor="account-name" className="block text-sm font-medium text-slate-200 mb-2">
-            Display Name
-          </label>
-          <input
-            id="account-name"
-            type="text"
-            value={user.name}
-            disabled
-            className="w-full px-4 py-2 rounded-xl border border-white/15 bg-cosmic-card-solid text-white disabled:opacity-60"
-          />
-        </div>
-
-        <div>
-          <label htmlFor="account-email" className="block text-sm font-medium text-slate-200 mb-2">
-            Email Address
-          </label>
-          <input
-            id="account-email"
-            type="email"
-            value={user.email}
-            disabled
-            className="w-full px-4 py-2 rounded-xl border border-white/15 bg-cosmic-card-solid text-white disabled:opacity-60"
-          />
-          <p className="mt-1 text-xs text-slate-200">
-            Email cannot be changed. Contact support for assistance.
-          </p>
-        </div>
-
-        <div>
-          <label htmlFor="account-timezone" className="block text-sm font-medium text-slate-200 mb-2">
-            Timezone
-          </label>
-          <input
-            id="account-timezone"
-            type="text"
-            value={user.timezone}
-            disabled
-            className="w-full px-4 py-2 rounded-xl border border-white/15 bg-cosmic-card-solid text-white disabled:opacity-60"
-          />
-        </div>
-
-        <div>
-          <label htmlFor="label-member-since" className="block text-sm font-medium text-slate-200 mb-2">
-            Member Since
-          </label>
-          <input
-            type="text"
-            id="label-member-since"
-            value={new Date(user.createdAt).toLocaleDateString(APP_LOCALE, {
-              month: 'long',
-              day: 'numeric',
-              year: 'numeric',
-            })}
-            disabled
-            className="w-full px-4 py-2 rounded-xl border border-white/15 bg-cosmic-card-solid text-white disabled:opacity-60"
-          />
-        </div>
-      </div>
-
-      {/* Change Password Section */}
-      <div className="pt-6 border-t border-white/15">
-        <h4 className="text-md font-medium text-white mb-4">Security</h4>
-        <button type="button" className="px-4 py-2 bg-primary text-white rounded-lg hover:bg-primary/90 transition-colors">
-          Change Password
-        </button>
-      </div>
-
-      {/* Danger Zone */}
-      <div className="pt-6 border-t border-white/15">
-        <h4 className="text-md font-medium text-red-400 mb-4">Danger Zone</h4>
-        <button type="button" className="px-4 py-2 border border-red-800 text-red-400 rounded-lg hover:bg-red-900/20 transition-colors">
-          Delete My Account
-        </button>
-        <p className="mt-2 text-xs text-slate-200">
-          This action cannot be undone. All your data will be permanently deleted.
-        </p>
-      </div>
-    </div>
-  );
-}
-
-// Charts Tab Component
-function ChartsTab({
-  charts,
-  onEditChart,
-  onViewChart,
-  onDeleteChart,
-}: {
-  charts?: Chart[];
-  onEditChart?: (chartId: string) => void;
-  onViewChart?: (chartId: string) => void;
-  onDeleteChart?: (chartId: string) => void;
-}) {
-  const navigate = useNavigate();
-  if (!charts || charts.length === 0) {
-    return (
-      <div className="text-center py-12">
-        <EmptyState
-          icon="auto_awesome"
-          title="No charts yet"
-          description="Create your first natal chart to get started"
-          actionText="Create Your First Chart"
-          onAction={() => navigate('/charts/new')}
-        />
-      </div>
-    );
-  }
-
-  return (
-    <div className="space-y-6" data-testid="charts-tab-content">
-      <div className="flex items-center justify-between">
-        <h3 className="text-lg font-semibold text-white">
-          My Charts ({charts.length})
-        </h3>
-        <button type="button" className="flex items-center gap-2 px-4 py-2 bg-primary text-white rounded-lg hover:bg-primary/90 transition-colors">
-          <span className="material-symbols-outlined text-4xl" aria-hidden="true">add</span>
-          Add New Chart
-        </button>
-      </div>
-
-      <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
-        {charts.map((chart) => (
-          <ChartCard
-            key={chart.id}
-            chart={chart}
-            onEdit={() => onEditChart?.(chart.id)}
-            onView={() => onViewChart?.(chart.id)}
-            onDelete={() => onDeleteChart?.(chart.id)}
-          />
-        ))}
-      </div>
-    </div>
-  );
-}
-
-// Chart Card Component
-function ChartCard({
-  chart,
-  onEdit,
-  onView,
-  onDelete,
-}: {
-  chart: Chart;
-  onEdit?: () => void;
-  onView?: () => void;
-  onDelete?: () => void;
-}) {
-  return (
-    <div className="bg-white/15 rounded-lg border border-white/15 overflow-hidden hover:shadow-lg transition-shadow">
-      {/* Chart Thumbnail */}
-      <div className="aspect-square bg-gradient-to-br from-white/5 to-white/[0.02] flex items-center justify-center p-4">
-        <div className="w-full h-full flex items-center justify-center text-slate-200">
-          <span className="text-sm">Chart Preview</span>
-        </div>
-      </div>
-
-      {/* Chart Info */}
-      <div className="p-4">
-        <h4 className="font-semibold text-white mb-1">{chart.name}</h4>
-        <p className="text-sm text-slate-200 mb-2">
-          {new Date(chart.birthData.date).toLocaleDateString(APP_LOCALE, {
-            month: 'long',
-            day: 'numeric',
-            year: 'numeric',
-          })}{' '}
-          • {chart.birthData.place.name}
-        </p>
-        <div className="flex items-center gap-2 text-xs text-slate-200">
-          <span className="capitalize">{chart.settings.houseSystem}</span>
-          <span>•</span>
-          <span className="capitalize">{chart.settings.zodiac}</span>
-        </div>
-      </div>
-
-      {/* Actions */}
-      <div className="px-4 pb-4 flex gap-2">
-        <button
-          type="button"
-          onClick={onView}
-          className="flex-1 px-3 py-2 bg-primary text-white text-sm font-medium rounded-lg hover:bg-primary/90 transition-colors"
-        >
-          View
-        </button>
-        <button
-          type="button"
-          onClick={onEdit}
-          className="px-3 py-2 border border-white/15 text-slate-200 text-sm font-medium rounded-lg hover:bg-white/15 transition-colors"
-          aria-label="Edit chart"
-        >
-          <span className="material-symbols-outlined text-4xl" aria-hidden="true">edit</span>
-        </button>
-        <button
-          type="button"
-          onClick={onDelete}
-          className="px-3 py-2 border border-red-800 text-red-400 text-sm font-medium rounded-lg hover:bg-red-900/20 transition-colors"
-          aria-label="Delete chart"
-        >
-          <span className="material-symbols-outlined text-4xl" aria-hidden="true">delete</span>
-        </button>
-      </div>
-    </div>
-  );
-}
-
-// Preferences Tab Component
-function PreferencesTab({ user }: { user?: UserProfile }) {
-  if (!user) return null;
-
-  return (
-    <div className="space-y-6">
-      <h3 className="text-lg font-semibold text-white">Chart Preferences</h3>
-
-      <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-        {/* Default House System */}
-        <div>
-          <label htmlFor="house-system" className="block text-sm font-medium text-slate-200 mb-2">
-            Default House System
-          </label>
-          <select
-            id="house-system"
-            defaultValue={user.preferences.defaultHouseSystem}
-            className="w-full px-4 py-2 rounded-lg border border-white/15 bg-white/15 text-white focus:border-primary focus:ring-primary"
-          >
-            <option value="placidus">Placidus</option>
-            <option value="koch">Koch</option>
-            <option value="porphyry">Porphyry</option>
-            <option value="whole">Whole Sign</option>
-            <option value="equal">Equal</option>
-            <option value="topocentric">Topocentric</option>
-          </select>
-        </div>
-
-        {/* Default Zodiac Type */}
-        <div>
-          <label htmlFor="zodiac-type" className="block text-sm font-medium text-slate-200 mb-2">
-            Default Zodiac Type
-          </label>
-          <select
-            id="zodiac-type"
-            defaultValue={user.preferences.defaultZodiac}
-            className="w-full px-4 py-2 rounded-lg border border-white/15 bg-white/15 text-white focus:border-primary focus:ring-primary"
-          >
-            <option value="tropical">Tropical</option>
-            <option value="sidereal">Sidereal</option>
-          </select>
-        </div>
-      </div>
-
-      {/* Aspect Orbs */}
-      <div className="pt-6 border-t border-white/15">
-        <h4 className="text-md font-medium text-white mb-4">Aspect Orb Sensitivity</h4>
-        <div className="space-y-4">
-          <div>
-            <div className="flex justify-between mb-2">
-              <label htmlFor="conjunction-orb" className="text-sm font-medium text-slate-200">
-                Conjunction/ Opposition
-              </label>
-              <span className="text-sm text-slate-200">
-                {user.preferences.aspectOrbs.conjunction}°
-              </span>
-            </div>
-            <input
-              id="conjunction-orb"
-              type="range"
-              min="1"
-              max="15"
-              defaultValue={user.preferences.aspectOrbs.conjunction}
-              className="w-full"
-            />
-          </div>
-          <div>
-            <div className="flex justify-between mb-2">
-              <label htmlFor="trine-orb" className="text-sm font-medium text-slate-200">Trine/ Square</label>
-              <span className="text-sm text-slate-200">
-                {user.preferences.aspectOrbs.trine}°
-              </span>
-            </div>
-            <input
-              id="trine-orb"
-              type="range"
-              min="1"
-              max="15"
-              defaultValue={user.preferences.aspectOrbs.trine}
-              className="w-full"
-            />
-          </div>
-          <div>
-            <div className="flex justify-between mb-2">
-              <label htmlFor="sextile-orb" className="text-sm font-medium text-slate-200">Sextile</label>
-              <span className="text-sm text-slate-200">
-                {user.preferences.aspectOrbs.sextile}°
-              </span>
-            </div>
-            <input
-              id="sextile-orb"
-              type="range"
-              min="1"
-              max="12"
-              defaultValue={user.preferences.aspectOrbs.sextile}
-              className="w-full"
-            />
-          </div>
-        </div>
-      </div>
-
-      {/* Theme */}
-      <div className="pt-6 border-t border-white/15">
-        <h4 className="text-md font-medium text-white mb-4">Appearance</h4>
-        <div>
-          <label htmlFor="label-theme-light" className="block text-sm font-medium text-slate-200 mb-2">
-            Theme
-          </label>
-          <div
-            role="radiogroup"
-            aria-label="Theme"
-            className="grid grid-cols-3 gap-4"
-            onKeyDown={(e) => {
-              const themes = ['light', 'dark', 'auto'];
-              const idx = themes.indexOf(user.preferences.theme);
-              if (e.key === 'ArrowRight' && idx < themes.length - 1) { e.preventDefault(); user.preferences.theme = themes[idx + 1] as 'auto' | 'light' | 'dark'; }
-              if (e.key === 'ArrowLeft' && idx > 0) { e.preventDefault(); user.preferences.theme = themes[idx - 1] as 'auto' | 'light' | 'dark'; }
-            }}
-          >
-            {[
-              { value: 'light', label: 'Light', icon: '☀️' },
-              { value: 'dark', label: 'Dark', icon: '🌙' },
-              { value: 'auto', label: 'System', icon: '💻' },
-            ].map((theme) => (
-              <button
-                type="button"
-                key={theme.value}
-                role="radio"
-                id={theme.value === 'light' ? 'label-theme-light' : undefined}
-                aria-checked={user.preferences.theme === theme.value}
-                onClick={() => {
-                  const root = document.documentElement;
-                  if (theme.value === 'light') {
-                    root.classList.remove('dark');
-                    root.classList.add('light');
-                  } else if (theme.value === 'dark') {
-                    root.classList.remove('light');
-                    root.classList.add('dark');
-                  } else {
-                    // auto — follow system
-                    root.classList.remove('light', 'dark');
-                    if (window.matchMedia('(prefers-color-scheme: dark)').matches) {
-                      root.classList.add('dark');
-                    } else {
-                      root.classList.add('light');
-                    }
-                  }
-                  localStorage.setItem('theme', theme.value);
-                }}
-                className={`
-                  p-4 rounded-lg border-2 text-center transition-all
-                  ${
-                    user.preferences.theme === theme.value
-                      ? 'border-primary bg-primary/10'
-                      : 'border-white/15 hover:border-cosmic-border'
-                  }
-                `}
-              >
-                <div className="text-2xl mb-2">{theme.icon}</div>
-                <div className="text-sm font-medium text-white">{theme.label}</div>
-              </button>
-            ))}
-          </div>
-        </div>
-      </div>
-
-      {/* Save Button */}
-      <div className="pt-6">
-        <button type="button" className="px-6 py-3 bg-primary text-white rounded-lg hover:bg-primary/90 transition-colors">
-          Save Preferences
-        </button>
-      </div>
-    </div>
-  );
-}
-
-// Subscription Tab Component
-function SubscriptionTab({ user }: { user?: UserProfile }) {
-  if (!user) return null;
-
-  const plans = [
-    {
-      name: 'Free',
-      price: '$0/month',
-      features: ['3 natal charts', 'Basic personality analysis', 'Daily transits', 'Community support'],
-      cta: 'Current Plan',
-      disabled: true,
-      highlight: false,
-    },
-    {
-      name: 'Premium',
-      price: '$9.99/month',
-      features: [
-        'Unlimited natal charts',
-        'Detailed personality analysis',
-        'Transit calendar & forecasts',
-        'Aspect pattern detection',
-        'Priority support',
-      ],
-      cta: user.subscription.plan === 'premium' ? 'Current Plan' : 'Upgrade',
-      disabled: user.subscription.plan === 'premium',
-      highlight: true,
-    },
-    {
-      name: 'Professional',
-      price: '$29.99/month',
-      features: [
-        'Everything in Premium',
-        'Synastry charts (compatibility)',
-        'Composite charts',
-        'Transit to transit analysis',
-        'PDF chart exports',
-        'Dedicated support',
-      ],
-      cta: user.subscription.plan === 'professional' ? 'Current Plan' : 'Upgrade',
-      disabled: user.subscription.plan === 'professional',
-      highlight: false,
-    },
-  ];
-
-  return (
-    <div className="space-y-6">
-      {/* Current Plan */}
-      <div className="bg-gradient-to-r from-primary/10 to-violet-500/10 rounded-lg p-6 border border-primary/20">
-        <div className="flex items-start justify-between">
-          <div>
-            <h3 className="text-lg font-semibold text-slate-200 mb-2">
-              Current Plan: {user.subscription.plan.charAt(0).toUpperCase() + user.subscription.plan.slice(1)}
-            </h3>
-            <p className="text-sm text-primary">
-              Status: {user.subscription.status.charAt(0).toUpperCase() + user.subscription.status.slice(1)}
-            </p>
-            {user.subscription.renewalDate && (
-              <p className="text-xs text-primary mt-1">
-                Renews on {new Date(user.subscription.renewalDate).toLocaleDateString(APP_LOCALE, {
-                  month: 'long',
-                  day: 'numeric',
-                  year: 'numeric',
-                })}
-              </p>
-            )}
-          </div>
-          <span className="material-symbols-outlined text-4xl text-primary" aria-hidden="true">auto_awesome</span>
-        </div>
-      </div>
-
-      {/* Available Plans */}
-      <div>
-        <h4 className="text-lg font-semibold text-white mb-4">Available Plans</h4>
-        <div className="grid grid-cols-1 md:grid-cols-3 gap-6">
-          {plans.map((plan) => (
-            <PlanCard key={plan.name} plan={plan} />
-          ))}
-        </div>
-      </div>
-
-      {/* Billing History */}
-      <div className="pt-6 border-t border-white/15">
-        <h4 className="text-lg font-semibold text-white mb-4">Billing History</h4>
-        <div className="bg-white/15 rounded-lg p-6 text-center">
-          <p className="text-slate-200 text-sm">
-            No billing history available for your account.
-          </p>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-// Plan Card Component
-function PlanCard({
-  plan,
-}: {
-  plan: {
-    name: string;
-    price: string;
-    features: string[];
-    cta: string;
-    disabled: boolean;
-    highlight: boolean;
-  };
-}) {
-  return (
-    <div
-      className={`
-        bg-white/15 rounded-lg border-2 p-6
-        ${plan.highlight ? 'border-primary shadow-lg' : 'border-white/15'}
-      `}
-    >
-      {plan.highlight && (
-        <div className="text-center mb-4">
-          <span className="px-3 py-1 bg-primary/10 text-primary text-xs font-medium rounded-full">
-            Most Popular
-          </span>
-        </div>
-      )}
-      <div className="text-center mb-6">
-        <h3 className="text-xl font-bold text-white mb-2">{plan.name}</h3>
-        <p className="text-3xl font-bold text-primary">{plan.price}</p>
-      </div>
-      <ul className="space-y-3 mb-6">
-        {plan.features.map((feature) => (
-          <li key={feature} className="flex items-start gap-2 text-sm text-slate-200">
-            <span className="material-symbols-outlined text-4xl text-green-500 flex-shrink-0 mt-0.5" aria-hidden="true">check</span>
-            {feature}
-          </li>
-        ))}
-      </ul>
-      <button
-        type="button"
-        disabled={plan.disabled}
-        className={`
-          w-full py-3 rounded-lg font-medium transition-colors
-          ${
-            plan.disabled
-              ? 'bg-white/15 text-slate-200 cursor-not-allowed'
-              : plan.highlight
-              ? 'bg-primary text-white hover:bg-primary/90'
-              : 'bg-white/15 border border-white/15 text-slate-200 hover:bg-white/15'
-          }
-        `}
-      >
-        {plan.cta}
-      </button>
     </div>
   );
 }
