+++ frontend/src/components/user-profile/index.ts
@@ -1,0 +1,16 @@
+export { ProfileHeader } from './ProfileHeader';
+export type { ProfileHeaderProps } from './ProfileHeader';
+
+export { AccountTab } from './AccountTab';
+export type { AccountTabProps } from './AccountTab';
+
+export { ChartsTab } from './ChartsTab';
+export type { ChartsTabProps } from './ChartsTab';
+
+export { PreferencesTab } from './PreferencesTab';
+export type { PreferencesTabProps } from './PreferencesTab';
+
+export { SubscriptionTab } from './SubscriptionTab';
+export type { SubscriptionTabProps } from './SubscriptionTab';
+
+export type { UserProfile, Chart, HouseSystem, UserProfileProps, UserData } from './types';
